android_reversing/requirements.txt
Flask
google-genai
markdown
python-dotenv
Flask-Cors 
PyYAML